chore: Using a subview for indicating loading
